Split BaseMeshGradientRenderer to future movement to commonMain
