+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/graphics/DefaultMeshGradientRenderer.skiko.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.graphics
+
+import androidx.compose.ui.geometry.Offset
+
+// TODO: Move to commonMain, it doesn't use skiko APIs.
+//  Android is different only because it avoids extra allocations by calling Android APIs directly.
+internal class DefaultMeshGradientRenderer : BaseMeshGradientRenderer() {
+    private val paint = Paint()
+
+    override fun drawTriangles(
+        canvas: Canvas,
+        surfacePositions: FloatArray,
+        surfaceColors: IntArray,
+        indices: ShortArray,
+        vertexCount: Int,
+    ) {
+        val vertexPositions = List(vertexCount) { i ->
+            Offset(surfacePositions[i * 2], surfacePositions[i * 2 + 1])
+        }
+        val vertexColors = List(vertexCount) { i -> Color(surfaceColors[i]) }
+        val vertexIndices = List(indices.size) { i -> indices[i].toInt() }
+
+        canvas.drawVertices(
+            vertices = Vertices(
+                vertexMode = VertexMode.Triangles,
+                positions = vertexPositions,
+                textureCoordinates = vertexPositions,
+                colors = vertexColors,
+                indices = vertexIndices,
+            ),
+            blendMode = BlendMode.Dst,
+            paint = paint,
+        )
+    }
+}
